Project Submission
This is an upgraded CSC 15 project from years ago that bothered me to
this day as my coding at the time was horrible. It is a simple grade
book project was allows a user to create a class, add students, alter
user scores, delete students from a class,  sort students, back up
students to a "Class Files" folder using the class name for the file and
load students from a file using the class name. The files the students
are being saved to and loaded from are 2007 Excel files using the Apache
POI Excel java jar files.